requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-login==0.6.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
anthropic==0.34.2
apscheduler==3.10.4
slack-sdk==3.27.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
newsapi-python==0.2.7
feedparser==6.0.11
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
itsdangerous==2.1.2
